moved tests to be colocated with components
