+++ client/src/App.test.jsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
-import App from '../src/App';
+import App from './App';
 
 describe('something truthy and falsy', () => {
   it('true to be true', () => {
