+++ client/src/components/ProductMain/ProductDetails/ProductDetails.test.jsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
-import ProductDetails from '../src/components/ProductMain/ProductDetails/ProductDetails';
+import ProductDetails from './ProductDetails';
 
 
 describe('ProductDetails', () => {
